--- src/backend/common/src/main/java/site/danjam/mate/common/exception/AccessDeniedException.java
@@ -1,7 +1,0 @@
-package site.danjam.mate.common.exception;
-
-public class AccessDeniedException extends BaseException {
-    public AccessDeniedException() {
-        super(Code.ACCESS_DENIED);
-    }
-}
