--- src/backend/common/src/main/java/site/danjam/mate/common/exception/CanNotFindUserException.java
@@ -1,11 +1,0 @@
-package site.danjam.mate.common.exception;
-
-public class CanNotFindUserException extends BaseException {
-    public CanNotFindUserException(String message) {
-        super(Code.CAN_NOT_FIND_RESOURCE, message);
-    }
-
-    public CanNotFindUserException(Code code, String message) {
-        super(code,message);
-    }
-}
