--- src/backend/common/src/main/java/site/danjam/mate/common/exception/ExpiredTokenException.java
@@ -1,7 +1,0 @@
-package site.danjam.mate.common.exception;
-
-public class ExpiredTokenException extends BaseException {
-    public ExpiredTokenException(Code code) {
-        super(code);
-    }
-}
