--- src/backend/common/src/main/java/site/danjam/mate/common/exception/MultipartException.java
@@ -1,7 +1,0 @@
-package site.danjam.mate.common.exception;
-
-public class MultipartException extends BaseException {
-    public MultipartException() {
-        super(Code.USER_INVALID_FILE_EXTENSION);
-    }
-}
