--- src/backend/common/src/main/java/site/danjam/mate/common/exception/NotFoundUserException.java
@@ -1,7 +1,0 @@
-package site.danjam.mate.common.exception;
-
-public class NotFoundUserException extends BaseException {
-    public NotFoundUserException() {
-        super(Code.USER_CAN_NOT_FIND_USER);
-    }
-}
