+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/gateway_service/CanNotFindTokenException.java
@@ -1,4 +1,7 @@
-package site.danjam.mate.common.exception;
+package site.danjam.mate.common.exception.gateway_service;
+
+import site.danjam.mate.common.exception.BaseException;
+import site.danjam.mate.common.exception.Code;
 
 public class CanNotFindTokenException extends BaseException {
     public CanNotFindTokenException(String message) {
