+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/mate_service/AlreadyProfileExistException.java
@@ -1,4 +1,7 @@
-package site.danjam.mate.common.exception;
+package site.danjam.mate.common.exception.mate_service;
+
+import site.danjam.mate.common.exception.BaseException;
+import site.danjam.mate.common.exception.Code;
 
 public class AlreadyProfileExistException extends BaseException {
     public AlreadyProfileExistException() {
