+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/user_service/DuplicateUsernameException.java
@@ -1,4 +1,4 @@
-package site.danjam.mate.user_service.domain.certification.exception;
+package site.danjam.mate.common.exception.user_service;
 
 
 import site.danjam.mate.common.exception.BaseException;
