+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/user_service/MultipartException.java
@@ -1,0 +1,10 @@
+package site.danjam.mate.common.exception.user_service;
+
+import site.danjam.mate.common.exception.BaseException;
+import site.danjam.mate.common.exception.Code;
+
+public class MultipartException extends BaseException {
+    public MultipartException() {
+        super(Code.USER_INVALID_FILE_EXTENSION);
+    }
+}
