+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/user_service/NotFoundMyProfileException.java
@@ -1,5 +1,8 @@
-package site.danjam.mate.common.exception;
+package site.danjam.mate.common.exception.user_service;
 
+
+import site.danjam.mate.common.exception.BaseException;
+import site.danjam.mate.common.exception.Code;
 
 public class NotFoundMyProfileException extends BaseException {
     public NotFoundMyProfileException() {
