+++ src/backend/common/src/main/java/site/danjam/mate/common/exception/user_service/RefreshTokenOwnerMismatchException.java
@@ -1,4 +1,7 @@
-package site.danjam.mate.common.exception;
+package site.danjam.mate.common.exception.user_service;
+
+import site.danjam.mate.common.exception.BaseException;
+import site.danjam.mate.common.exception.Code;
 
 public class RefreshTokenOwnerMismatchException extends BaseException {
     public RefreshTokenOwnerMismatchException(String message) {
