+++ src/backend/user-service/src/main/java/site/danjam/mate/user_service/auth/service/RefreshTokenService.java
@@ -8,12 +8,12 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import site.danjam.mate.common.annotation.MethodDescription;
 import site.danjam.mate.common.exception.Code;
-import site.danjam.mate.common.exception.ExpiredTokenException;
-import site.danjam.mate.common.exception.InvalidTokenException;
+import site.danjam.mate.common.exception.global.ExpiredTokenException;
+import site.danjam.mate.common.exception.global.InvalidTokenException;
 import site.danjam.mate.user_service.auth.domain.RefreshToken;
 import site.danjam.mate.user_service.auth.security.JWTUtil;
 import site.danjam.mate.user_service.auth.repository.RefreshTokenRepository;
-import site.danjam.mate.common.exception.RequiredArgumentException;
+import site.danjam.mate.common.exception.global.RequiredArgumentException;
 
 @Service
 @RequiredArgsConstructor
