+++ src/backend/user-service/src/main/java/site/danjam/mate/user_service/domain/school/repository/MajorRepository.java
@@ -4,10 +4,9 @@ import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import site.danjam.mate.common.exception.CanNotFindResourceException;
+import site.danjam.mate.common.exception.global.CanNotFindResourceException;
 import site.danjam.mate.common.exception.Code;
 import site.danjam.mate.user_service.domain.school.domain.Major;
-import site.danjam.mate.user_service.domain.school.domain.School;
 
 @Repository
 @RequiredArgsConstructor
