+++ src/backend/user-service/src/main/java/site/danjam/mate/user_service/domain/user/repository/UserRepository.java
@@ -1,12 +1,11 @@
 package site.danjam.mate.user_service.domain.user.repository;
 
-import java.util.List;
 import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import site.danjam.mate.user_service.domain.user.domain.User;
 
-import site.danjam.mate.common.exception.NotFoundMyProfileException;
+import site.danjam.mate.common.exception.user_service.NotFoundMyProfileException;
 import site.danjam.mate.user_service.domain.certification.domain.Certification;
 
 @Repository
